+++ src/index.ts
@@ -1,8 +1,8 @@
 import { config } from './utils/config';
-import { downloadCacheAndFillDb } from './utils/database';
+import { downloadCacheAndFillDb } from './modules/iptv';
 import { getLogger } from './utils/logger';
 import { client } from './utils/discord';
-import { initializeStreamer } from './utils/discord_stream';
+import { initializeStreamer } from './modules/streaming';
 
 const logger = getLogger();
