--- src/modules/commands.ts
@@ -1,4 +1,0 @@
-export { handleStreamCommand } from './commands/stream';
-export { handleStopCommand } from './commands/stop';
-export { handleListCommand } from './commands/list';
-export { handleRefreshCommand } from './commands/refresh';
