+++ src/modules/commands/index.ts
@@ -1,0 +1,5 @@
+export { handleStreamCommand } from './stream';
+export { handleStopCommand } from './stop';
+export { handleListCommand } from './list';
+export { handleRefreshCommand } from './refresh';
+export { handleProgrammeCommand } from './programme';
